Update: Adds the payment service

+++ src/main/java/org/example/dto/UserDto.java
@@ -4,14 +4,14 @@ import lombok.Data;
 
 @Data
 public class UserDto {
-    private Integer id;
+    private Long id;
     private String phoneNumber;
     private String createdAt;
     private String name;
     private String email;
     private String orgName;
 
-    public UserDto(final Integer id, final String phoneNumber, final String createdAt, final String name, final String email, final String orgName) {
+    public UserDto(final Long id, final String phoneNumber, final String createdAt, final String name, final String email, final String orgName) {
         this.id = id;
         this.phoneNumber = phoneNumber;
         this.createdAt = createdAt;
